azure-functions/embedding-converter/requirements.txt
azure-functions
openai
numpy
python-dotenv
